fix: address the dtype, wording and Bokeh points from #57
- input_conversion indexed with `np.cumsum(x)` in the `rr_ms` branch, so a
  float array -- the natural dtype for intervals in milliseconds -- raised
  "arrays used as indices must be of integer (or boolean) type". The `rr_s`
  branch immediately below already cast for exactly this reason. Every
  input/output combination now works for both int and float.
- "X must be a 1darray" reads as "1d array" in the four places it is raised.
- Document that Bokeh's own output_notebook() must be called once per session
  before figures render in a notebook. Nothing is shown without it, which is
  not obvious from the Systole side.

The CDSView deprecation also raised in that issue was fixed when the Bokeh
backend was migrated to the 3.x API.

+++ systole/utils.py
@@ -663,11 +663,14 @@ def input_conversion(
             if output_type == "rr_s":
                 output = x / 1000
             elif output_type == "peaks":
+                # Intervals in milliseconds are naturally floats, and a float
+                # array cannot be used to index. The `rr_s` branch below already
+                # casts for the same reason.
                 output = np.zeros(int(np.sum(x)) + 1, dtype=bool)
-                output[np.cumsum(x)] = True
+                output[np.cumsum(x).astype(int)] = True
                 output[0] = True
             elif output_type == "peaks_idx":
-                output = np.cumsum(x)
+                output = np.cumsum(x).astype(int)
                 output = np.insert(output, 0, 0)
         else:
             raise ValueError("Invalid intervals provided.")
